# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EduConnect EdTech Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (super admin / admin / planner)
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [student, admin, super_admin, planner]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, active]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

    post:
      operationId: createAdmin
      tags: [users]
      summary: Create admin user (super admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminBody"
      responses:
        "201":
          description: Admin created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (super admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /users/{id}/approve:
    patch:
      operationId: approveStudent
      tags: [users]
      summary: Approve or reject a student registration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveStudentBody"
      responses:
        "200":
          description: Student status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /whiteboard/{classId}:
    get:
      operationId: getWhiteboardData
      tags: [whiteboard]
      summary: Get whiteboard state for a class
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Whiteboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhiteboardData"

    post:
      operationId: saveWhiteboardData
      tags: [whiteboard]
      summary: Save whiteboard state
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveWhiteboardBody"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhiteboardData"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    CreateAdminBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        fullName:
          type: string
        email:
          type: string
        subject:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
          enum: [admin, planner]
      required:
        - username
        - password
        - fullName
        - email
    UpdateUserBody:
      type: object
      properties:
        fullName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
    ApproveStudentBody:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected]
      required:
        - status
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, admin, super_admin, planner]
        status:
          type: string
          enum: [pending, approved, rejected, active]
        phone:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - fullName
        - email
        - role
        - status
        - createdAt
...
    WhiteboardData:
      type: object
      properties:
        id:
          type: integer
        classId:
          type: integer
        data:
          type: string
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - classId
        - data
        - updatedAt
    SaveWhiteboardBody:
      type: object
      properties:
        data:
          type: string
      required:
        - data
...